tests/plugin_tests/ui/website/django/requirements.txt
djangorestframework
markdown
django-filter
pygments
django-cors-headers